Re-work of how smoke grenade projectile actors play smoke sounds & handle their own destruction: Got rid of using Tick to do a timed actor destruction - there's no need to use Tick, simply set a LifeSpan when it explodes and the actor will auto-destroy after that time. On a server I tear off the actor when it explodes and set a LifeSpan of 1 second - just to make sure bTearOff reaches each client. After that the actor isn't needed on a server & gets destroyed. On a client, the smoke grenade plays the loud, hissing smoke ignition sound as a one off when it explodes (about 4,5 secs). Then carries on playing the smoke loop sound for as long as the specified SmokeSoundDuration, using a simple Timer. I have found 33 seconds to be just right - you can see & hear it clearly on the coloured smokes, with sound stopping when they stop discharging & smoke begins to drift away. When the Timer is called it shuts off the smoke sound & sets a LifeSpan of 10 secs, meaning the empty grenade lies on the ground for a bit, then the actor destroys itself.
Can do something similar with tank & mortar smoke shells, but need to think about what sound they would make - would it be similar to a grenade?

NOTE - all smoke projectiles need to be bAlwaysRelevant, to make sure the smoke effect is spawned on each client. Smoke effects are big and last for a while, so it's quite possible the projectile wouldn't otherwise be considered net relevant to a client, so it would never get the smoke effect, then the player comes into range/sight/relevancy of the smoke cloud - but they don't see it (cue accusations of cheating). So have to make 'em bAlwaysRelevant. Once effect is spawned on client, actor can be torn off and then destroyed on server though, so only has short server/replication life.