Re-worked the new DH_Bullet a little more: 1) Removed tracer properties (mostly light) from defprops & now only set them in PostNetBeginPlay() if it is a tracer bullet. 2) Removed use of Tick() to match rotation of tracer to bullet's velocity, as I found a handy actor variable called bOrientToVelocity that does that automatically & natively. 3) Refactored a little, mostly to avoid possible accessed none errors.
I may revert to having a separate tracer bullet class, to avoid setting the tracer props when it spawns. I combined normal & tracer bullets into one class partly due to simplicy of that, but also so it isn't necessary to copy the speed etc properties from the non-tracer bullet into the defprops of the tracer version.

Don't think it makes much difference either way. By far the overriding advantage is that tracer bullets are now real bullets with visual effects added and we don't have separate dummy tracer bullets spawned on the client as well as the real bullet received through replication.