Remove RO's vehicle repeating 1 sec Timer() that ran on every vehicle until: (1) a driver had entered (which setsbDriverAlreadyEntered, allowing vehicle to be damaged by firendly fire, i.e. it protects against idiots griefing in spawn); and (2) the vehicle had crew (deprecated in DH).
bDriverAlreadyEntered is instead set in KDriverEnter() or, in single player (where protection is not required), in PostBeginPlay().
PostBeginPlay() is modified to avoid setting an initial timer.
Timer() is stripped down.

This avoids loads of undriven vehicles constantly running a timer and checks.
But more importantly it paves the way for using the vehicle's timer for something more useful, where we hugely strip back vehicle Tick().