Fix for https://bitbucket.org/darklightgames/darkesthour/issue/141/no-exhaust-effect-when-vehicle-replicated & is empty but with engine running
Problem is because Tick() is disabled when vehicle is empty & Tick is where the emitters are updated.
Exhaust emitters are being spawned in StartEmitters(), but without a Tick() update, they spawn no particles.
It's necessary to call a nil velocity UpdateExhaust() on the ExhaustEffect, which sets particle emissions to an engine idle level.

Solution is to call this in StartEmitters(), although it's only necessary if the vehicle is not being driven (because if it is then Tick will be running).