Class DH_PlaceableWeaponPickup - re-worked substantially so now is a generic, placeable actor that can be used as a pickup for any weapon type. Was an abstract parent class for satchel & faust pickups. There was also placeable PIAT pickup, which I had renamed as it originally had a very confusing name. But I have now deleted it, as it is rather obscure & no longer needed anyway.
I thought the placeable PIAT pickup had never been used, but I have found it in the non-official Pegasus Bridge map.
When I checked there were 2 actual PIAT pickups, but also another 5 that had been changed to give rifles or MG42s.
The leveller had changed the static mesh, weapon & ammo classes, but it only worked properly in single player.
In multi-player, very little is replicated, including the StaticMesh property, which normally would but a Pickup class has bOnlyReplicateHidden, which stops that.

This got me thinking & I've ended re-writing the DH_PlaceableWeaponPickup class & turning it into a non-abstract, placeable actor.
A leveller can use this where they want to add specific weapon pickups, just setting the WeaponType class in the editor, without the need for lots of weapon-specific subclasses.
The only replication I have added is the WeaponType class & it only replicates once and only if it has been set in the editor & doesn't match the default value (so subclasses still don't do any replication).
Fixed bugs where it would allow you to pick up & carry more then one weapon in the same inventory slot (e.g. 2 primary weapons).
Made it so respawning (on a timer, set to say 9999 for a one-time pickup) & match reset both work.
Disabled all replication when pickup has been used & re-enable if it respawns or match is reset (can't destroy actor, even if not set to respawn as needs to be there if match reset).
Had to add to message classes for touch & pickup messages, as need to pass them a reference to the weapon class, so they can access it's name in def props.
Made it so the correct StaticMesh, DrawScale & InventoryType (ammo) are set automatically from the weapon class, as its own PickupClass & FireModeClass have those properties.
A smart leveller will at least set the mesh in the editor, so it looks right in editor and can be accurately positioned.
Only bit that doesn't seem to work properly is if you already have the weapon & are below full ammo, so you should pick up some ammo. Works for PIAT, satchel, faust, etc, but not for guns needing mags. Minor.
Overall a lot going on and some smart stuff - for a class that will probably barely be used ! But that's just me ........

ps - Pegasus Bridge map won't run in 6.0 anyway as it doesn't have a DH_LevelInfo. I have remade a V2 with a level info & with the offending PIAT pickups replaced with this new generic class. All works fine.