Rename vehicle classes:
DHTreadCraft to DHArmoredVehicle - may or may not have treads, so the real defining feature of this class is that the vehicle has 'proper' armour (i.e. may stop shell) - is also expected to have a turret/cannon, so probably the best descriptve name would be DHArmoredFightingVehicle, but that's too long. DHArmoredVehicle does the trick.

DHArmoredWheeledVehicle to DHArmoredCar - does what it says on the tin.

DHTransportCraft to DHApcVehicle - again, does what it says on the tin and nice & short too - defining features of this class are treads & light armour that will stop small arms - technically I suppose an APC could have wheels, but all the ones in game use treads so this is a reasonable & descriptive class name (TransportCraft implied trucks would be subclasses).

(DHWheeledVehicle stays as is).