Should be a full fix for https://bitbucket.org/darklightgames/darkesthour/issue/342/when-vehicles-replicate-the-correct vehicle & player animations often do not play. Been something of a marathon epic ! So many issues ...... Previous SetPlayerPosition() clientside function in various vehicle classes solves the problem most of the time.
But I've found that the Driver actor is another one of those vital actors that arrive on a net client in a random order through replication, adding to the holy trinity of vehicle, VWPawn and vehwep.
Sometimes a VWPawn has it's VehicleBase and Gun, but doesn't yet have the Driver actor - the player pawn.
But the delayed Driver replication problem actually affects vehicles more, because a VWPawn already waits until it receives its Gun actor, and that delay is usually enough to coincidentally have received the Driver too - but not always.
A vehicle doesn't have to wait for a Gun actor to replicate, so SetPlayerPosition() was being called from PostNetBeginPlay, as soon as the actor replicated.
That often meant it hadn't received it's Driver yet, so the functionality failed.
The solution is to rely on the vehicle having bDriving=true, which is reliable even if the Driver actor hasn't replicated yet.
On a net client, if bDriving is true in PostNetBeginPlay, we now set a new bNeedToInitializeDriver flag.
Then in PostNetReceive, we look for the Driver arriving if bNeedToInitializeDriver is set.
That becomes the trigger to do any driver-related setup - in this case to call SetPlayerPosition() to play the correct anims on the vehicle & driver.
Added same to cannons & MGs, because although they usually work, it's only a coincidence because of the delay waiting for Gun, and sometimes it can go wrong.

Also, to stop the problem of the DHPawn's StartDriving event overriding the correct Driver anim with the default DriveAnim, I now use bClientPlayedDriveAnim & bClientSkipDriveAnim in DHPawn.
StartDriving records that it has played DriveAnim on a net client at least once, so the vehicle knows that it isnt going to get overridden.
But if bClientPlayedDriveAnim is false and the vehicle plays a different Driver anim in SetPlayerPosition(), the vehicle flags bClientSkipDriveAnim on the DHPawn.
That stops the initial DriveAnim from playing, and is then reset so it has no further effect.

Finally, where a VWPawn's SetPlayerPosition() detaches & re-attaches the Driver, it now only does so if the DHPawn's StartDriving has flagged bNeedToAttachDriver.
DHPawn does that if it failed to attach the driver, because the VWPawn didn't  have its Gun actor.

ps - The block of apparent variable changes in DHPawn are just layout. There are only the 3 new bools added.