Fix for https://bitbucket.org/darklightgames/darkesthour/issue/293/halftrack-passengers-are-unusually susceptible to explosive damage from outside the vehicle. Problem is fundamental to UT2004 & new/re-worked functionality is added to handle blast damage to vehicle occupants.
Explosive projectiles call HurtRadius() in the proj calss, which does a foreach iteration to find all VisibleCollidingActors within blast radius.
Damage is reduced pro-rata based on distance from explosion.
But vehicle occupants don't have collision - it gets turned off when entering - so HurtRadius() doesn't find them.
To handle that,  it calls DriverRadiusDamage() on any Vehicle within the blast radius.
DriverRadiusDamage() damages the vehicle's driver or occupant of a VWPawn, if they are exposed & their location is within the blast radius.
Actually, it only does that if the occupant has no collision, otherwise HurtRadius() will get them anyway, but occupants don't have collision.
This system works, except it takes account of the occupant possibly being shielded from the blast by parts of the vehicle.

To fix this, when HurtRadius() finds a Vehicle, it now calls CheckVehicleOccupantsRadiusDamage() in the proj class, instead of DriverRadiusDamage() on the Vehicle.
DriverRadiusDamage() is deprecated.
CheckVehicleOccupantsRadiusDamage() checks for any exposed occupants and calls VehicleOccupantRadiusDamage() for each one.
VehicleOccupantRadiusDamage() is where the business gets done.
It does a foreach TraceActors iteration, from the blast centre to the occupant's head (specifically a point 3 inches below the top of their head).
Any Trace won't actually hit the occupant pawn, as it has no collision.
But this 'trace' is in reverse - it's checking whether anything is in the way, so a negative result means positive damage.
The trace continues until it hits a blocking actor, which usually means part of the Vehicle, in which case no damage is done to the occupant.
Or it continues to the end of the trace, meaning nothing blocked it so the occupant is damaged.

A foreach iteration may raise eyebrows, but TraceActors is pretty quick, like any trace, especially as it's a very short trace, which is abandoned if it hits any blocker.
A normal Trace would be stopped by the player's bullet whit attachment, which has collision but doesn't block.
It's possible to look for such actors in the bTrace result and then re-Trace from the new HitLocation, but that can go on and it gets messy.

And a big advantage of doing the 'Trace' in the proj class, is that any Trace function uses the collision properties of the actor being Traced from.
In this case it uses the projectile collision properties, specifically including bUseCollisionStaticMesh, so it checks collision accurately against a vehicle's collision static mesh.
Where as if it were done in the vehicle class, say in DriverRadiusDamage(), it would use the vehicle's collision props,
meaning not using bUseCollisionStaticMesh & inaccurate collision vs the blocky vehicle collision box.
That is a significant flaw in a number of areas, although I will shortly commit a fix for the most obvious.