Fix for DHBullet part of https://bitbucket.org/darklightgames/darkesthour/issues/345/passengers-can-be-shot-through-the-sides of some infantry vehicles. Does NOT fix the similar problems caused by bullet's pre-launch trace, which will be committed separately. Means that a spawned bullet's ProcessTouch() function, which is called when the bullet collides with a player's surrounding bullet whip attachment, will run a 2nd verification trace to make sure there isn't a blocking actor in the way, including a part of the vehicle. 2nd trace is a TraceActors iteration, allowing it to ignore any invalid traced actors (any that ProcessTouch would normally ignore) and continue tracing. Although it's a foreach iteration, it's very fast as TraceActors is a fast iteration, like any trace, and it's a one-off, on demand usage, with a trace over a very short distance, which exits if it hits any blocker.
Also a fix for DHBullet part of https://bitbucket.org/darklightgames/darkesthour/issues/341/cant-damage-exposed-vehicle-occupants-who are 'enveloped' in a collision box.
But again, does NOT fix the similar problems caused by bullet's pre-launch trace, which will be committed separately.
Now temporarily changes the Instigator's collision settings to bUseCollisionStaticMesh, as the trace is called on the Instigator pawn (the firing player) and all traces use the calling actor's collision settings.
Projectiles have bUseCollisionStaticMesh=true, but a player pawn does not, so before the change the Instigator's trace was ignoring vehicle's detailed collision meshes & just using the crude collision boxes.
The Instigator is reset to its default bUseCollisionStaticMesh setting as soon as the traces are done, so it's only for a split second.
And that is harmless anyway, because if we set a player pawn to bUseCollisionStaticMesh, all that means is it has more complex collision detection with vehicles, i.e. the player's feet would follow the hull's shape more closely when standing on the vehicle, rather than simply standing on the crude collision box.