Improvement & simplification of system for using collision static mesh actors (CSMA) attached to VehicleWeapons like tank turrets, which can be used to give them accurate hit detection, like a vehicle hull. Remember a col static mesh added to a VW mesh in the editor doesn't work, because it doesn't rotate with the turret's yaw bone. The main change - and simplification - is to make the CSMA an ordinary static mesh actor (extends StaticMeshActor). Much easier and cleaner than original method, which required an 'empty', invisible skeletal mesh that used the col static mesh, which was then attached to the turret's yaw bone. A static mesh actor can do precision, per triangle collision detection, and can 'pass on' any hit to the rear VW actor, the same as the old method.
There is no replication here - the CSMA is spawned independently on the server and clients.
Projectile's run their own hit detection calcs against local actors (the server is definitive only re any resulting damage).

As before, the VW removes all of its own collision if it spawns a CSMA, and the CSMA copies the collision of the real VW actor (in effect, it becomes its collision proxy).
Now added CSMA checks to the bullet pre-launch trace system, and also to the 2nd, verification trace that I recently added to DHBullet.
Another substantial simplification is that the need for separate subclasses for each vehicle's CSMA has been removed.
Instead of specifying a particular CSMA class, we now simply specify the col static mesh to use, i.e. a static mesh file and mesh name.
The system spawns the generic CSMA class and just sets its static mesh after it spawns.

New system also automatically adjusts the relative offset position of the CSMA to take account of the yaw bone's position, so need need to pre-calc offsets or adjust the col mesh model.
Also automatically hides the CSMA, so no need to skin it invisble and it can be seen in the editor, which is more user-friendly.
I have updated the notes on how to use a CSMA and it's shorter and easier now.

Added a useful debug exec 'ShowColMesh' to cannon and MG pawn, which 1st toggles shows the col mesh as a white shape alongside the VW, then to hide the VW mesh, then returns to normal.
Works in multi-player as it's only making visual changes on the client. Like most debug tools, it needs debug mode or single player to operate.

Have renamed the CSMA to DHCollisionStaticMeshActor instead of DHVehicleWeaponCollisionMeshActor, because it's shorter and because I can see potential use for things other than VehicleWeapons.

Have dusted down my old tiger turret collision mesh. Removed the commander, as not required with new vehicle occupant hit detection we're using now (actually gets in the way).
Also removed all concavity in the model by splitting it into separate, 'closed' mesh parts.
Made new col meshes for MGs with gunshields - CSMAs for German halftrack and StuG/StuH's remote MG.
Also made a col mesh for early StuG's fixed gunshield, but that one doesn't need to be a CSMA, as it doesn't rotate and so can use a normal col mesh set in the VW skeletal mesh in editor.
Will upload these separately, as they are linked to some other mesh file changes. Probably tomorrow.