Slight mod to DHHud.DrawPlayerNames() to fic accessed none error (think it was on PawnOwner.PRI.Team). Extra use of GetTeamNum() function. Save to local variable to avoid running the it twice.
Also removed bit about spectators being able to see names of all players (when PawnOwner.PRI.Team = none or GetTeamNum = 255).
It doesn't actually happen (testing in 5.1 and RO and is same). Spectators don't get this function called.

Updated comments about resupplying mortars to be more accurate.
Slight re-factor to avoid cast to OtherDHP if we now we're looking at a deployed mortar, not a player.
(This makes it look like a lot of changes, but it isn't; it's just a block being indented 1 tab right).