Fix for https://bitbucket.org/darklightgames/darkesthour/issues/352/vehicle-commanders-in-binoculars-position don't actually hold binocs (just grasp invisible ones). Spawns clientside binocs deco attachment & attached to commander's hands. Destroyed when move away from binocs or exit vehicle.
Also a fix for https://bitbucket.org/darklightgames/darkesthour/issues/370/in-many-vehicles-commander-moving-down from binoculars plays a wierd 'arm waving' animation.
Calls Driver.SetAnimFrame(1.0) when playing an anim on moving down from binocs, which jumps to the end of the anim & freezes there, so it effectively become an unbuttoned idle anim, which is what we need.