Players will no longer spawn at exit positions of spawn vehicles marked as SVT_Always. This stops players from inexorably being left in the dust of full and moving vehicles.
Spawn vehicles are now marked as being blocked if there are no non-crew positions available to be occupied. This isn't ideal, of course, as we'd like to fully support crews spawning inside of tanks that might be marked as spawn vehicles. Unfortunately this would create a lot more code for a use case that we don't *really* intend to support.

Fixed bug where players would witness their character die continuously if attempting to spawn at a spawn vehicle that had became blocked.