Update that stops the confusing use of SpawnCount  - SpawnCount in GRI is no longer used  - SpawnCount in DHGame is no longer used  - SpawnsRemaining is now used by everything (that I could find)
This change makes it much easier to grasp how reinforcements work.  This should also result in less code in the future.  Because changing the reinforcements was actually changing how many deaths occurred (which seems funky).  We can now have a separate variable to count actual casualties incurred, that we can later use for statistical purposes.