Updated ListPlayers command to also copy the info into the clipboard, so you can type the command and then paste the info, pretty useful to get ROIDs in the server via console
People often struggle to get their ROID, so I've created a way for the player to copy their ROID to their clipboard.  This requires a stupid hack because GetPlayerIDHash() does not work on the client end.  Because it requires the player to be connected to a server, I've made a new string on the client's end that is updated when a player connects to a dedicated server, this new string can be easily copied into the clipboard via the Settings page in the network area.  If the player hasn't yet connected to a server, the setting page informs the user, that they must first play multiplayer.