Assumed fix for https://bitbucket.org/darklightgames/darkesthour/issues/430/infantry-uniforms-can-be-seen-skinned
This is a complicated affair that I will try to explain.  First I should note that I may not be correct on this assumption.
Our pawns in DH did not have skin slots assigned for both the skin and the face, and our Heer pawns had the incorrect slot assigned for their skin.  RO pawns did have these things assigned, even if redundant.  Also an interesting note: One of the Axis radio pawns had no skin assigned (axis radiomen are often messed up texture wise, in 5.1 and 6.0).  The skin slots in the pawn are overrode by the UPL file, but the UPL override seems to have a chance of failure (both in 5.1 and 6.0).  Our new spawning system seems to make these UPL failures happen more often, this is why the issue is more noticeable in 6.0.

Why is the UPL failing more often?
I'm not sure, but I believe it has something to do with spawning in the black box and being immediately teleported to the spawn point.

So why does having these skins fix the issue?
It does not fix the UPL failing, however it is assigning correct default skins for when the UPL does fail.

Why does it not use the mesh's default skins?
I don't have a clue, and it is very dumb that it doesn't.

So basically these skin slots assigned in the pawn are responsible for being the base texture assignment.  I believe this to actually be a re-discovery as I believe I discovered this working on Darkest Orchestra.  Testing this fix was not 100% certain, this is why I'm still just assuming all of this.  First, I had to test with bots, as it is too hard to get lots of players.  Second, the bots don't seem to have the issue permanently like players do, this could be somewhere between better hardware, better network, or lack of a player net environment.  Regardless, I  did notice for less than a second or so the texture would be messed up when spawning the bots.  After assigning the texture slots, I noticed to not see that and thus I'm assuming I've fixed the issue.

In regards to the UPL system and its failing, after 6.0, we will do work to remove the UPL system and to have the random assignment of skins and faces to be done via code and to have UPLs not allowed at all (completely removed).  It is too late for now to attempt to do this.