1) Moved `simulated function SpawnExplosionEffects()` from DHCannonShell.uc to DHAntiVehicleProjectile.uc (included required variables) ^ This should not cause any bugs, because the function is new and only existed in DHCannonShell.uc (and thus is only called on that class)
2) EST_Custom00 material type is now a used material type in DH!
It prevents emitters, sounds, and decals from spawning/playing with DHBullets, Tracers, and ATProjectiles (Mortars may not be required, but we can add them if it becomes an issue)

This is to allow for hidden BSP that I use in levels to cut network to not expose bullets, tracers, etc in mid air (which looks very strange).

^ Why not just use a new actor and prevent effects from ever trying to play and even perhaps let the projectile continue it's course?
BSP is world geometry the HitWall function sends LevelInfo as what it hit.  This complicates the matter and forces us to rely on material.

Note: I've set the material type on the solid colors in DH_Engine_tex.utx to be surface type EST_Custom00 and that will require SVN update to see this effectual.