- Re-enabled PLT (Prelaunch Trace on DHProjectileFire) - Adjusted the BulletWhipAttachement to not be twice as large (changed back to RO default, which we believe 5.1 was actually using)
Note: This may actually not have a large impact on CPU performance, however it may help when all those bullets are flying in CQC.  Really the purpose of this change is for the hit detection issues we have seen since 6.0 and we hope this will restore the nice hit detection DH seemed to have in the past.  If we do notice a restoration of CPU performance, then it is likely we can further improve upon the CPU performance by dynamically extending out the PLT trace range when the server is stressed.  The bullet whip attachment size reduction might actually have a larger benefit to the CPU than PLT.  Because each time a bullet enters one of these it has to do checks, the more the slower the CPU cycle, and thus lowering the tick rate.  The reduction is probably for the best anyways, because I've noticed myself getting blur from rounds outside a bunker (while I was inside it).