Fix for https://github.com/DarklightGames/DarkestHour/issues/503
The following shows the track Left and Right number, found in Vehicle_Engines.uax and UC is Universal Carrier tracks found in Vehicle_EnginesTwo
Allied Tanks:
07 : Cromwells
UC : All others

Axis Tanks:
03 : Small (PIII, Marder)
05 : Medium (Panther, PIVs)
04 : Big (Tiger, KT, JagdPanther, JagdTiger)
08 : TD/AG (JagdPanzer, StugIII, StuH42)

Note: I've reduced the track sounds overall by 10%, so they can only be 90% as loud as the engine.  But it should actually always be 10% less than what the engine is playing at, so it is a significant change.  I've also slightly increased the default sound radius (600 to 650), this will make the vehicles a bit louder and heard from further away.