Should be a fix for https://github.com/DarklightGames/DarkestHour/issues/381 (Bullet pre-launch trace does not play any bullet whiz/snap sounds). Also https://github.com/DarklightGames/DarkestHour/issues/521 (Player is protected behind breakable glass). Will write up a proper explanation of this tomrrow, but brief note3s for this commit:
Takes out a fair bit of pointless variable replication re whizzes (mWhizSoundLocation & SpawnWhizCount in DHPawn).
Also removes a fair bit of pointless calculation (BulletDistance & WhizType in script, WhizLocation & calling PawnWhizzed in native) on server and clients.
Stops whiz being called at all on a server in bullet collision (use WhizType 0 in HitPointTrace), as server does not play whizzes or replicate them to clients.
In bullet collision, each client runs its own HitPointTrace and calculates and receives whizzes accordingly - but have stopped all but the whizzed player's client from doing this, as is pointless for others.

PLT now uses special WhizType 255 on server, which makes native calculate WhizLocation & call PawnWhizzed, but naturally the server doesn't play whizzes.
Instead it just grabs the WhizLocation and saves the whizzed player, whiz type and whiz location in a local array.
Then at the end of PLT, if it is successful and is going to stop a bullet from spawning, server loops through array of saved whizzes and calls a replicated function on each client, making it call PawnWhizzed and play the correct whiz.