Add more useful vehicle debug options: ** NEEDS SVN UPDATE ** LogVehDamage() LogVehAttach() SetHUDTreads(int NewPosX0, int NewPosX1, int NewPosY, int NewScale) SetMass(float NewValue) SetTreadHeight(float NewValue) DebugAngles(optional string Option, optional float NewAngle)
Note the last 2 use temporary static mesh debug attachments to show planes.
Here they are used to give a clear visual indication of the current highest point where a side hit is counted as a tread hit,
and how the 4 corner angles are set to calculate front/left/rear/right side hits.
(Which are the cause of virtually all strange penetration problems in the game and many false accusations of cheating - but that's a whole different story .....)
To avoid the DH_DebugTools static mesh file being always loaded, I use `DynamicLoadObject("DH_DebugTools.Misc.DebugPlaneAttachment", class'StaticMesh')` to load on demand.