Sd.Kfz.234/1 cannon - couple of minor updates to def props:
Removed bIsRepeatingFF=true, which I previously added with comment "strongly suspect this should be true, like a tank mounted MG" (this is just force feedback).
Now I've spent time tracking the intricacies of the vehicle weapon fire/cease fire mechanisms, I realise this is incorrect.
Although the 20mm autocannon is similar to an MG in many ways, it plays individual fire effects for each round fired, where an MG plays a continuous, ambient fire effects until you cease fire.

And updated AIInfo (just missed from last commit, with general AIInfo clean up). RefireRate sounds like a rate of repeating fire setting, but it isn't - it's the probability of a bot firing again, expressed as a float between 0 and 1, and with auto fire weapons it should be set to 0.00.