Yay - a fix for https://github.com/DarklightGames/DarkestHour/issues/353 (firing MG34 in single-shot mode while hip-firing results in significant firing offset).
Problem was on the server, affecting both real projectiles & triggered hit effects from the bullet pre-launch trace system.
Server had the wrong muzzle position & direction in the 1st person weapon model.
Although when you hip the MG34, you see it swing up, no animation plays on the server.
So when the server fired from the hip, it fired from the lowered position, with the muzzle much further left and down from where you saw it on your screen.
And it fired diagonally to the left, where it believed the weapon was pointing.
Realising that's what the server was doing quickly led to finding the cause & the solution.

So why only this weapon?
Well, typical weapons only have 2 positions: hipped and ironsighted (bipod deployed is the same as ironsighted really).
As weapon raising/lowering animations don't play on the server, for typical weapons it is set permanently in the hip position, which is the starting/neutral position.
So that's fine for hip firing typical weapons.
When you ironsight, the server is completely at odds with where the weapon really is.
But that doesn't matter, beacuse in ironsights or bipod deployed modes, the server uses you eye location as the base to calculate the bullet's position & direction.
It doesn't need the muzzle location or direction for ironsights, so all is well, despite the server being out of synch.
An MG weapon has, or can have, THREE positions: lowered, hipped & bipod deployed.
But the MG34 is the only MG that allows hip firing, so it's the only one that could be affected.

So why only single shot mode and why was it fixed by an auto burst?
That was the randomising factor, which made it a bit tricky to debug!
For no good reason, the only weapon animation that plays on the server is for auto firing.
It doesn't need to, as the anim only shows little effects like recoil, bolt, etc.
In fact, I'm going to stop the server playing those anims, as they are pointless and just waste processing resources.
But in this case it meant that as soon as the MG34 fired auto, the server played an animation that snapped the weapon model to the correct hip firing position on the server.
Even firing auto from bipod did the same, as that still snapped the weapon up from its lowered, carried position and pointed it forward.
So auto fire worked fire, and once it had been auto fired, the server's weapon model was placed in the hip fire position - and left there.

The fix: On a server a weapon plays the 'idle' animation (for the resting position) once when the weapon spawns, which snaps a standard weapon to the hip fire position & leaves it there.
In the MG class I've modified this, so that if the MG has a hip fire mode, the server plays an idle animation for the hip fire position instead of the resting/lowered position.
Meaning the MG34 weapon model on the server will always be in the hip fire position & that's all it needs.