ATTENTION!  Important, please read:
Okay so I've removed the /System/ file from the ignore list, this is the RO/System folder!!!

I've added ignore checks to ignore as many files as I thought might exist in this folder, but I'm sure you may have more files than what I have.

You can exclude additional files in a variety of ways, but the exclude file will only do it on your local end, this way our .gitignore file doesn't get out of hand.

To add a file to the exclude list, when you go to commit, right-click it and select the "Add to ignore list" if you can do the file type (then prob just add it to the .gitignore), but specific files (like .ini files) you'll need to exclude!
Select the file's name then check: "Only in the containing folders" and ".git/info/exclude"

I've done this so I can have DHMapVotingInfo.ini in the RO/System folder as to my knowledge this is the only location it'll work.

Added DHMapVotingInfo.ini with all official maps and values that I thought were best.