Fix for bug https://github.com/DarklightGames/DarkestHour/issues/537
This is a slight partial rewrite of how mortar targets are stored and cleared.

I suspect the *actual* cause of this was that a player who changed teams would have `ClearMortarTarget` called *after* the team was switched. This means that the team & index look-up would be erroneous, resulting in bizarre behavior. To fix this, instead of using a team and target index to clear targets, the controller (`DHPlayer`) is passed into `ClearMortarTarget`, and mortar targets are cleared if they have a matching controller. This provides a more robust solution.

Removed `MortarTargetIndex`, a variable that provided a negligible performance benefit when clearing targets. For some reason it was a replicated variable, so now that's less data to replicate, also.