Fix for bug https://github.com/DarklightGames/DarkestHour/issues/592.
Fixes the case where a user is a tank crewman with a vehicle selected, then selects a non-tank crewman role and is then seemingly unable to spawn until the user manually selects "None" from the vehicle list.

A pop-up message will now be displayed when the user has a tank selected to spawn in the deploy menu and it becomes disabled for any reason. It will also select the "None" option in this case.

There is now no situation where the Continue button should be greyed out without an obvious reason (eg. no spawn selected).